+++ sshmitm/moduleparser/__init__.py
@@ -1,8 +1,9 @@
-from sshmitm.moduleparser.modules import BaseModule, SubCommand
+from sshmitm.moduleparser.modules import BaseModule, SSHMITMBaseModule, SubCommand
 from sshmitm.moduleparser.parser import ModuleParser
 
 __all__ = [
     "BaseModule",
     "ModuleParser",
+    "SSHMITMBaseModule",
     "SubCommand",
 ]